requirements.txt
pydantic==1.10.11
torch==2.0.1
torchvision==0.15.2
dependency-injector==4.41.0
fastapi==0.103.1
python-multipart==0.0.5
scikit-learn==1.3.0
uvicorn==0.23.2
Pillow==10.0.0
timm==0.9.2
omegaconf==2.3.0
numpy==1.24.3
opencv-python==4.8.1.78
wemake-python-styleguide==0.18.0